# Speed up Profile Deletion checks