--- a/core/src/modules/ops/profile.ts
+++ b/core/src/modules/ops/profile.ts
@@ -671,26 +671,37 @@ export namespace ProfileOps {
       (await plugin.readSetting("core", "runs-profile-batch-size")).value
     );
 
-    const profiles: Profile[] = await api.sequelize.query(
-      `
-SELECT "id" FROM "profiles"
-WHERE "state"='ready'
-AND 0 = (
-  SELECT count("exports"."id") FROM "exports"
-    WHERE "exports"."profileId"="profiles"."id"
-    AND "exports"."state" IN ('pending', 'processing')
-)
-AND "id" NOT IN (
-  SELECT DISTINCT("profileId") FROM "profileProperties"
-  JOIN properties ON "properties"."id"="profileProperties"."propertyId"
-  WHERE "properties"."directlyMapped"=true AND "rawValue" IS NOT NULL
-)
-LIMIT ${limit};
-      `,
-      {
-        model: Profile,
-      }
+    let profiles: Profile[] = [];
+    const directlyMappedProperties = (await Property.findAllWithCache()).filter(
+      (p) => p.directlyMapped
     );
+
+    if (directlyMappedProperties.length === 0) {
+      // We have no directly mapped Property and every profile should be removed
+      profiles = await Profile.findAll({ attributes: ["id"], limit });
+    } else {
+      // We have directly mapped Properties and we only want to remove those Profiles with a null "user_id" (directlyMapped) Property
+      profiles = await api.sequelize.query(
+        `
+  SELECT "id" FROM "profiles"
+  WHERE "state"='ready'
+  AND 0 = (
+    SELECT count("exports"."id") FROM "exports"
+      WHERE "exports"."profileId"="profiles"."id"
+      AND "exports"."state" IN ('pending', 'processing')
+  )
+  AND "id" NOT IN (
+    SELECT DISTINCT("profileId") FROM "profileProperties"
+    JOIN properties ON "properties"."id"="profileProperties"."propertyId"
+    WHERE "properties"."directlyMapped"=true AND "rawValue" IS NOT NULL
+  )
+  LIMIT ${limit};
+        `,
+        {
+          model: Profile,
+        }
+      );
+    }
 
     return profiles;
   }