Using:
```
./bin/timed_run staging-enterprise 100
```

Timed Run Before this PR:
```
🎉 Complete in 0hrs 55min 30sec [3330 seconds]
```

Timed Run After this PR:
```
🎉 Complete in 0hrs 48min 35sec [2915 seconds]
```

Not much faster locally...